services/llm-service/pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "pydantic-settings>=2.3",
    "structlog>=24.1",
    "openai>=1.30",
    "chromadb==0.5.23",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_service"]
